.claude/skills/zimage-core/requirements.txt
# Z-Image Core Dependencies
replicate>=0.20.0
Pillow>=10.0.0
requests>=2.28.0
python-dotenv>=1.0.0
